% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{list_cfx}
\alias{list_cfx}
\title{List the CFX csv datasets found in a folder.}
\usage{
list_cfx(folder)
}
\arguments{
\item{folder}{a character vector specifying the path(s) to the folder containing the cfx datasets.}
}
\value{
A vector with the basenames (cfx_set) to the different datasets
}
\description{
List the CFX csv datasets found in a folder.
}

