% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_amplification}
\alias{read_amplification}
\title{Load the RFU amplification values}
\usage{
read_amplification(cfx_run)
}
\arguments{
\item{cfx_run}{a string with the path to the CFX run information file or a folder containing a single run information file.}
}
\value{
A dataframe containing the RFU values kinetic
}
\description{
Load the RFU amplification values
}

