% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_cq}
\alias{read_cq}
\title{Load the Cq values}
\usage{
read_cq(cfx_run)
}
\arguments{
\item{cfx_run}{a string with the path to the CFX run information file or a folder containing a single run information file.}
}
\value{
A dataframe containing the Cq values
}
\description{
Load the Cq values
}

