% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_meltcurve}
\alias{read_meltcurve}
\title{Load the derivative of the melt curve RFU values kinetic}
\usage{
read_meltcurve(cfx_set)
}
\arguments{
\item{cfx_set}{a string basename combining the folder and the common prefix of exported csv files. Use list_cfx to list the different sets in a folder.}
}
\value{
A dataframe containing the melt curve RFU derivative values
}
\description{
Load the derivative of the melt curve RFU values kinetic
}

