% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_meltpeak}
\alias{read_meltpeak}
\title{Load the melt cruve peaks calculated by the CFX Manager software}
\usage{
read_meltpeak(cfx_run)
}
\arguments{
\item{cfx_run}{a string with the path to the CFX run information file or a folder containing a single run information file.}
}
\value{
A dataframe containing the peak informations
}
\description{
Load the melt cruve peaks calculated by the CFX Manager software
}

