% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_amplification}
\alias{read_amplification}
\title{Load the RFU amplification values}
\usage{
read_amplification(cfx_set)
}
\arguments{
\item{cfx_set}{a string basename combining the folder and the common prefix of exported csv files. Use list_cfx to list the different sets in a folder.}
}
\value{
A dataframe containing the RFU values kinetic
}
\description{
Load the RFU amplification values
}

