% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_cq}
\alias{read_cq}
\title{Load the Cq values}
\usage{
read_cq(cfx_set)
}
\arguments{
\item{cfx_set}{a string basename combining the folder and the common prefix of exported csv files. Use list_cfx to list the different sets in a folder.}
}
\value{
A dataframe containing the Cq values
}
\description{
Load the Cq values
}

