% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_meltcurve}
\alias{read_meltcurve}
\title{Load the derivative of the melt curve RFU values kinetic}
\usage{
read_meltcurve(cfx_run)
}
\arguments{
\item{cfx_run}{a string with the path to the CFX run information file or a folder containing a single run information file.}
}
\value{
A dataframe containing the melt curve RFU derivative values
}
\description{
Load the derivative of the melt curve RFU values kinetic
}

