% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfx.R
\name{read_meltpeak}
\alias{read_meltpeak}
\title{Load the melt cruve peaks calculated by the CFX Manager software}
\usage{
read_meltpeak(cfx_set)
}
\arguments{
\item{cfx_set}{a string basename combining the folder and the common prefix of exported csv files. Use list_cfx to list the different sets in a folder.}
}
\value{
A dataframe containing the peak informations
}
\description{
Load the melt cruve peaks calculated by the CFX Manager software
}

